requirements.txt
streamlit==1.24.0
pandas==2.0.3
numpy==1.24.3
groq==0.4.0
python-dotenv==1.0.0
openpyxl==3.1.2
matplotlib==3.7.2
plotly==5.15.0
python-docx==0.8.11